refactor : 입력 받는 view 에서 유효성 체크 하는 유틸 별도 분리
입력 받는 view 에서 유효성 체크 하는 유틸 별도 분리
